# golang-api/assets/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "title": "Golang API.",
    "version": "0.0.1-alpha"
  },
  "paths": {
...
    "/ratings": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Ratings"
        ],
        "summary": "Retrieves a paginated ratings of movies.",
        "operationId": "ListAllMovieRatings",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Page",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Limit",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseListAllMovieRatings"
          },
          "500": {
            "$ref": "#/responses/GenericErrorResponse"
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Ratings"
        ],
        "summary": "Adds a new ratings of a movie to the system.",
        "operationId": "AddRating",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "genres": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "x-go-name": "Genres"
                },
                "id": {
                  "type": "string",
                  "x-go-name": "ID"
                },
                "original_language": {
                  "type": "string",
                  "x-go-name": "OriginalLanguage"
                },
                "popularity": {
                  "type": "string",
                  "x-go-name": "Popularity"
                },
                "release_date": {
                  "type": "string",
                  "x-go-name": "ReleaseDate"
                },
                "runtime": {
                  "type": "string",
                  "x-go-name": "Runtime"
                },
                "spoken_languages": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "x-go-name": "SpokenLanguages"
                },
                "status": {
                  "type": "string",
                  "x-go-name": "Status"
                },
                "title": {
                  "type": "string",
                  "x-go-name": "Title"
                }
              }
            }
          }
        ]
      }
    },
    "/ratings/movies/{movieId}/ratings": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Ratings"
        ],
        "summary": "Retrieves a ratings of movie by ID.",
        "operationId": "GetRatingsByMovieId",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "MovieID",
            "name": "movieId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseGetRatingsByMovieId"
          },
          "500": {
            "$ref": "#/responses/GenericErrorResponse"
          }
        }
      }
    },
    "/ratings/movies/{movieId}/user/{userId}/ratings": {
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Ratings"
        ],
        "summary": "Updates an existing rating.",
        "operationId": "UpdateRating",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "MovieID",
            "name": "movieId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "UserID",
            "name": "userId",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "movieId": {
                  "type": "string",
                  "x-go-name": "MovieId"
                },
                "rating": {
                  "type": "string",
                  "x-go-name": "Rating"
                },
                "userId": {
                  "type": "string",
                  "x-go-name": "UserId"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GenericSuccessResponse"
          },
          "400": {
            "$ref": "#/responses/ValidationErrorResponse"
          },
          "500": {
            "$ref": "#/responses/GenericErrorResponse"
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Ratings"
        ],
        "summary": "Deletes rating of a movie from the system.",
        "operationId": "DeleteRating",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "MovieID",
            "name": "movieId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "UserID",
            "name": "userId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GenericSuccessResponse"
          },
          "400": {
            "$ref": "#/responses/GenericErrorResponse"
          },
          "500": {
            "$ref": "#/responses/GenericErrorResponse"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Ratings": {
      "type": "object",
      "properties": {
        "movieId": {
          "type": "string",
          "x-go-name": "MovieId"
        },
        "rating": {
          "type": "string",
          "x-go-name": "Rating"
        },
        "userId": {
          "type": "string",
          "x-go-name": "UserId"
        }
      },
      "x-go-package": "git.pride.improwised.dev/Onboarding-2025/Krupanshi-Vaishnav/go-api/models"
    }
  },
  "responses": {
    "GenericErrorResponse": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "status": {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Status"
          }
        }
      }
    },
    "GenericSuccessResponse": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "status": {
            "type": "integer",
            "format": "int64",
            "enum": [
              "success"
            ],
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseGetRatingsByMovieId": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/definitions/Ratings"
          },
          "status": {
            "type": "string",
            "enum": [
              "success"
            ],
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ValidationErrorResponse": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "status": {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Status"
          }
        }
      }
    }
  }
}